Criados alguns endpoints para chamado

+++ api/src/main/java/br/ufpr/api/repository/ChamadoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.ufpr.api.model.entity.Chamado;
 
-public interface ChamadoRepository extends CrudRepository<Chamado, Long> {
+public interface ChamadoRepository extends CrudRepository<Chamado, Integer> {
     public Chamado findBySlug(String slug);
     List<Chamado> findAllByOrderByDataCriacaoAsc();
